dummy commit to go back

+++ lib/db/functions.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_video_info/flutter_video_info.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:permission_handler/permission_handler.dart';
@@ -17,7 +18,7 @@ ValueNotifier<List> filteredFolderVideos =
     ValueNotifier([]); //folder click videos
 ValueNotifier<List<String>> favVideos = ValueNotifier([]);
 
-ValueNotifier<List<RecentModel>> recentVideos = ValueNotifier([]);
+List<RecentModel> recentVideos = [];
 
 onSuccess(List<String> data) {
   fetchedVideosPath = data;
@@ -97,10 +98,11 @@ getFolderVideos(String path) {
 //video info collection
 
 String formatTime(double time) {
-    Duration duration = Duration(milliseconds: time.round());
-    return [duration.inHours, duration.inMinutes, duration.inSeconds].map((seg) => seg.remainder(60).toString().padLeft(2, '0')).join(':');
+  Duration duration = Duration(milliseconds: time.round());
+  return [duration.inHours, duration.inMinutes, duration.inSeconds]
+      .map((seg) => seg.remainder(60).toString().padLeft(2, '0'))
+      .join(':');
 }
-
 
 Future getVideoWithInfo() async {
   fetchedVideosWithInfo.value.clear();
@@ -145,26 +147,25 @@ sortByDate() {
 
 fetchFav() async {
   // if (box.get('favList') != null) {//first case there is no fav list then hive.get is an error
-    favVideos.value.clear();
-    // List temp = await box.get('favList');
-    // for (int i = 0; i < temp.length; i++) {
-    //   //delete video from fav if the user delete/change directory of the video in storage
-    //   if (!fetchedVideosPath.contains(temp[i])) {
-    //     print("found");
-    //     temp.remove(temp[i]);
-    //   }
-    // }
-    // favVideos.value.addAll(temp);
-    // favVideos.notifyListeners();
-    // await box.put('favList', temp);
-    List<String> temp = await favDB.get('favList').favVideo;
-    favVideos.value.addAll(temp);
-    print(favDB.get('favList').favVideo);
-
-  }
+  favVideos.value.clear();
+  // List temp = await box.get('favList');
+  // for (int i = 0; i < temp.length; i++) {
+  //   //delete video from fav if the user delete/change directory of the video in storage
+  //   if (!fetchedVideosPath.contains(temp[i])) {
+  //     print("found");
+  //     temp.remove(temp[i]);
+  //   }
+  // }
+  // favVideos.value.addAll(temp);
+  // favVideos.notifyListeners();
+  // await box.put('favList', temp);
+  List<String> temp = await favDB.get('favList').favVideo;
+  favVideos.value.addAll(temp);
+  print(favDB.get('favList').favVideo);
+}
 // }
 
-addToFavList(String value) async{
+addToFavList(String value) async {
   favVideos.value.add(value);
   favVideos.notifyListeners();
   var fav = Favourites(favVideo: favVideos.value);
@@ -181,63 +182,30 @@ removeFromFav(String value) {
 
 //###################...recent section...####################
 
-getRecentList() async{
-  
-  
-  // values.addAll(await recentDB.values);
-  // List keys = recentDB.keys;
-  // recentDB.delete(6);
-  // print(values.toList());
-  // print(keys);
-  // recentVideos.value.clear();
-  if(recentDB.values != null){
-    List<RecentModel> dbRecent = await recentDB.values.toList();
-    print(dbRecent[0].recentDate);
-    recentVideos.value.addAll(dbRecent);
-    recentVideos.value.reversed;
-    recentVideos.notifyListeners();
-    // // print(recentVideos.value);
+getRecentList() async {
+  recentVideos.clear();
+  if (recentDB.values != null) {
+    List<RecentModel> dbRecent = recentDB.values.toList();
+    recentVideos.addAll(dbRecent);
+    recentVideos.reversed;
+    print("##################################################");
+    print(recentVideos.length);
   }
 }
 
-addToRecent(RecentModel value) async{
-  final keys = await recentDB.keys.toList();
-  final values = await recentDB.values.toList();
-  for(int i=0; i<keys.length; i++){
-    if(values[i].recentPath == value.recentPath){
-      await recentDB.delete(keys[i]);
-      break;
+addToRecent(RecentModel value) async {
+  if (recentDB.values.isNotEmpty) {
+    final keys = recentDB.keys.toList();
+    final values = recentDB.values.toList();
+    for (int i = 0; i < keys.length; i++) {
+      if (values[i].recentPath == value.recentPath) {
+        await recentDB.delete(keys[i]);
+        break;
+      }
     }
-  }  
-  await recentDB.add(value);
-  recentVideos.value.removeWhere((element) => element.recentPath == value.recentPath);//remove the path received form current recent list
-  recentVideos.value.insert(0,value);
-
-
-  // await recentDB.add(value);
-  // recentVideos.value.clear();
-  // recentVideos.value.addAll(await recentDB.values.toList());
-  // recentVideos.value.reversed;
-  // print(recentVideos.value);
-  // if(recentDB.values != null){
-  // deleteItem(value.recentPath);
-  // }
-  // // recentVideos.value.
-  // // var existingItem = items.firstWhere((itemToCheck) => itemToCheck.link == favoriteitem.link, orElse: () => null);
-  // recentVideos.value.add(value);
-  // getRecentList();
-  // // print(recentVideos.value);
-  // // recentDB.addAll();
-
-}
-
-deleteItem(String currentPath) {
-    // final Map<dynamic, RecentModel> deliveriesMap = recentDB.toMap();
-    // dynamic desiredKey;
-    // deliveriesMap.forEach((key, value){
-    //     if (value.recentPath == currentPath){//db path == adding path
-    //       desiredKey = key;
-    //     }
-    // });
-    // recentDB.delete(desiredKey);
   }
+  print(recentDB.values);
+  await recentDB.add(value);
+  getRecentList();
+}
+// clear(){print(recentDB.length);}
